% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{make_explorer_app}
\alias{make_explorer_app}
\title{Create Minc Model Explorer App}
\usage{
make_explorer_app(title, pydpiper_dir, out_dir, metadata, defs,
  model_formula = Filenames ~ genotype,
  config_file = system.file("config.R", package = "mincModelExplorer"),
  blur_level = 0.2, input_file_col = "Filenames", relevels = c(genotype =
  "WT"), plot_choices = "genotype", label_data = NULL,
  study_average = NULL, mask = NULL, clobber = FALSE)
}
\arguments{
\item{title}{The name of the app}

\item{out_dir}{Where to create the app directory}

\item{metadata}{A file containing the data needed to fit models of
absolute and relative jacobians. Can include a labels column indicating
MAGeT labelled data files}

\item{defs}{A definitions file}

\item{model_formula}{The basic model to run}

\item{blur_level}{For filtering pydpiper output}

\item{relevels}{A named character vector of column_name = base_factor_level
defaults to c("genotype" = "WT")}

\item{plot_choices}{values to offer the use for plot choices}

\item{label_data}{A file containing paths to MAGeT labelled data files}

\item{study_average}{Where to find the final average, defaults to the nlin-3 average
from pydpiper}

\item{mask}{A mask to use for the analysis, defaults to the nlin-3 mask}

\item{clobber}{whether to overwrite files in an existing app directory}
}
\value{
The final merged data frame invisibly.
}
\description{
Read in pydpiper files and directories and produce an appropriate app directory
}
